successful refactoring! tr and node cmd work
