+++ events/interactionCreate.js
@@ -1,43 +1,5 @@
 const { InteractionType } = require('discord.js');
-
-// storing default languages
-const Database = require("@replit/database");
-const db = new Database();
-
-// embed
-const { translateEmbed } = require('../modules/embeds.js');
-
-// npm i @vitalets/google-translate-api
-const translate = require('@vitalets/google-translate-api');
-
-// supported languages
-const langs = {
-  'ar': 'arabic',
-  'zh-CN': 'chinese (simplified)',
-  'zh-TW': 'chinese (traditional)',
-  'de': 'german',
-  'el': 'greek',
-  'en': 'english',
-  'es': 'spanish',
-  'fa': 'persian',
-  'fi': 'finnish',
-  'fr': 'french',
-  'he': 'hebrew',
-  'hi': 'hindi',
-  'it': 'italian',
-  'he': 'hebrew',
-  'ja': 'japanese',
-  'ko': 'korean',
-  'la': 'latin',
-  'nl': 'dutch',
-  'pt': 'portuguese',
-  'ru': 'russian',
-  'th': 'thai',
-  'tr': 'turkish',
-  'uk': 'ukrainian',
-  'vi': 'vietnamese',
-};
-
+const tr = require('../modules/tr'); // include the tr function
 
 // EXECUTE COMMANDS DYNAMICALLY
 module.exports = {
@@ -62,7 +24,7 @@ module.exports = {
     // language autocomplete
     else if (interaction.type === InteractionType.ApplicationCommandAutocomplete) {
       const focusedValue = interaction.options.getFocused();
-      const choices = Object.values(langs);
+      const choices = Object.values(tr.langs);
       const filtered = choices.filter(choice => choice.startsWith(focusedValue));
       await interaction.respond(
         filtered.map(choice => ({ name: choice, value: choice })),
@@ -71,24 +33,12 @@ module.exports = {
 
     // for context menu translations
     else if (interaction.isMessageContextMenuCommand()) {
+      // open 15-minute window
       await interaction.deferReply({ ephemeral: true });
-
-      // try user language first
-      let to = await db.get(interaction.user.id);
-      if (!to) {
-        // then try server language
-        to = await db.get(interaction.guild.id);
-        if (!to) {
-          // if all else fails, default to english
-          to = 'english';
-        }
-      }
-
-      // results
+      // words to translate
       const words = interaction.targetMessage.content;
-      const translated = await translate(words, { to: to });
-      const embed = translateEmbed(words, 'auto', to);
-      return await interaction.editReply({ content: translated.text, embeds: [embed] });
+      // reply with results
+      return await interaction.editReply(tr(words, from, to));
     }
   },
 };
